# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: NebulaCloud API
  version: "0.0.1"
  description: |
    NebulaCloud is a self-hosted PaaS in the spirit of Railway, Render, Heroku
    and Vercel. This document describes the public REST + WebSocket surface
    consumed by the dashboard and by automation clients.

    Phase 0 shipped health and meta. Phase 1 adds Identity. Phase 2 adds
    organizations, projects, services, environment variables, deployments,
    GitHub webhooks, and scaffolded logs/metrics queries for the dashboard.

    Subsequent phases will deepen Domains, streaming logs, and richer metrics.

  contact:
    name: NebulaCloud
    url: https://github.com/nebulacloud/nebula
servers:
  - url: http://api.nebula.localhost
    description: local development
  - url: https://api.example.com
    description: production (replace with your domain)
...
paths:
...
  /api/v1/deployments/{deploymentID}:
    get:
      tags: [Deployments]
      operationId: getDeployment
      summary: Get one deployment joined with project/service labels
      security:
        - BearerAuth: []
      parameters:
        - name: deploymentID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deployment.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Deployment" }
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeploymentActor:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
    Deployment:
      type: object
      required: [id, service_id, service_name, project_id, project_name, trigger, status, created_at]
      properties:
        id: { type: string, format: uuid }
        service_id: { type: string, format: uuid }
        service_name: { type: string }
        project_id: { type: string, format: uuid }
        project_name: { type: string }
        trigger: { type: string }
        status: { type: string }
        commit_sha: { type: string }
        commit_message: { type: string }
        ref: { type: string }
        image_ref: { type: string }
        error_message: { type: string }
        duration_ms: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        triggered_by: { $ref: "#/components/schemas/DeploymentActor" }
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [kind, message]
          properties:
            kind:
              type: string
              enum:
                - validation
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - unavailable
                - internal
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: { type: string }